requirements.txt
Django==2.2.16
django-debug-toolbar==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
Faker==12.0.1
gunicorn==20.0.4
mixer==7.1.2
Pillow==8.3.1
psycopg2-binary==2.8.6
python-dotenv
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.3.1
webcolors==1.11.1